+++ packages/docs-shared/lib/WordCount/createWordCountInfo.ts
@@ -1,18 +1,33 @@
-import { isCJK, isKana, isThai, isWordDelimiter, toCodePoint } from './CharacterUtils'
+import {
+  isCJK,
+  isKana,
+  isThai,
+  isWhitespace,
+  isWordCountSupported,
+  isWordDelimiter,
+  toCodePoint,
+} from './CharacterUtils'
 import type { WordCountInfo } from './WordCountTypes'
 
-const segmenter = new Intl.Segmenter('en', { granularity: 'grapheme' })
+const segmenter = isWordCountSupported ? new Intl.Segmenter('en', { granularity: 'grapheme' }) : undefined
 
 export const createWordCountInfo = (textContent: string): WordCountInfo => {
   let characterCount = 0
   let wordCount = 0
   let nonWhitespaceCharacterCount = 0
+
+  if (!segmenter) {
+    return { characterCount, wordCount, nonWhitespaceCharacterCount }
+  }
+
   let alreadyMatched = false
   for (const segmentData of segmenter.segment(textContent)) {
     const char = segmentData.segment
     const charCodePoint = toCodePoint(char)
 
-    if (char.trim() === char) {
+    const charIsWhitespace = isWhitespace(char)
+
+    if (!charIsWhitespace) {
       nonWhitespaceCharacterCount++
     }
 
@@ -22,7 +37,7 @@ export const createWordCountInfo = (textContent: string): WordCountInfo => {
       break
     }
 
-    const isMatched = !isWordDelimiter(charCodePoint)
+    const isMatched = !charIsWhitespace && !isWordDelimiter(charCodePoint)
     const shouldIncrement = !alreadyMatched || isCJK(charCodePoint) || isKana(charCodePoint) || isThai(charCodePoint)
 
     if (shouldIncrement && isMatched) {